fix: final implementation to produce results

+++ src/main/java/com/github/jlong4bc/weatherhistoryapi/NoaaWeather.java
@@ -15,7 +15,7 @@ import java.time.LocalDate;
 public class NoaaWeather
 {
     private LocalDate date;
-    private int highTemp;
-    private int lowTemp;
-    private int precipitationAmount;
+    private double highTemp;
+    private double lowTemp;
+    private double precipitationAmount;
 }
