+++ src/main/java/com/github/jlong4bc/weatherhistoryapi/elements/Precipitation.java
@@ -3,6 +3,6 @@ package com.github.jlong4bc.weatherhistoryapi.elements;
 /**
  * A record to group items of precipitation
  */
-public record Precipitation(int amount, PrecipitationUoM unitOfMeasure, PrecipitationType type)
+public record Precipitation(double amount, PrecipitationUoM unitOfMeasure, PrecipitationType type)
 {
 }
