+++ src/main/java/com/github/jlong4bc/weatherhistoryapi/elements/Temperature.java
@@ -4,6 +4,6 @@ package com.github.jlong4bc.weatherhistoryapi.elements;
 /**
  * A record to group all attributes of temperature
  */
-public record Temperature(int high, int low, TemperatureUoM unitOfMeasure)
+public record Temperature(double high, double low, TemperatureUoM unitOfMeasure)
 {
 }
